Somehow fixed Stationary attacks
It had to do with use_gravity putting the sprites (of the same size) one pixel into the ground when they switched.

Fixed by removing the bottom pixel from the attack sprite collision masks.

Then it started floating.

Put the pixel back in the mask. And voila now its working as intended.